pyproject.toml
[project]
name = "django-chess"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chess",
    "daphne>=4.2.1",
    "django>=5.2.5",
    "django-debug-toolbar>=6.0.0",
    "djangorestframework>=3.14.0",
    "django-cors-headers>=4.3.0",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=5.2.2",
    "mypy>=1.17.1",
    "pytest-django>=4.11.1",
]

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "django_chess.dev_settings"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "django_chess.dev_settings"

[tool.uv.sources]
chess = { git = "https://github.com/niklasf/python-chess.git", rev = "ffa04827e325de5b4d39a67eee3528474b814285" }
